Remove usage of env in favor of CLI options.

+++ packages/terra-functional-testing/src/services/wdio-selenium-docker-service.js
@@ -22,16 +22,19 @@ class SeleniumDockerService {
    */
   async onPrepare(config) {
     const { launcherOptions } = config;
-    const { keepAliveSeleniumDockerService } = launcherOptions || {};
+    const { disableSeleniumService, keepAliveSeleniumDockerService } = launcherOptions || {};
 
+    this.disableSeleniumService = disableSeleniumService === true;
     this.keepAliveSeleniumDockerService = keepAliveSeleniumDockerService === true;
 
-    try {
-      // Verify docker is installed.
-      await exec('docker -v');
-      await this.startSeleniumHub();
-    } catch (error) {
-      throw new SevereServiceError(error);
+    if (!this.disableSeleniumService) {
+      try {
+        // Verify docker is installed.
+        await exec('docker -v');
+        await this.startSeleniumHub();
+      } catch (error) {
+        throw new SevereServiceError(error);
+      }
     }
   }
 
@@ -117,7 +120,7 @@ class SeleniumDockerService {
     // When multiple test sessions are executed sequentially as specified in the WDIO script in the package.json file, the keepAliveSeleniumDockerService cli option
     // is used to indicate not to remove the currently deployed docker stack upon test completion as it will be used again for the next test session.
     // The docker stack is expected to be removed by the last test session when no keepAliveSeleniumDockerService cli option is specified.
-    if (!this.keepAliveSeleniumDockerService) {
+    if (!this.keepAliveSeleniumDockerService && !this.disableSeleniumService) {
       logger.info('Shutting down the docker selenium hub...');
 
       await exec(`docker-compose -f ${this.getDockerComposeFilePath()} down`);
@@ -126,4 +129,3 @@ class SeleniumDockerService {
 }
 
 module.exports = SeleniumDockerService;
-
